+++ app/src/main/java/de/baumann/browser/browser/BrowserController.java
@@ -1,9 +1,11 @@
 package de.baumann.browser.browser;
 
 import android.net.Uri;
+import android.os.Message;
 import android.view.View;
 import android.webkit.ValueCallback;
 import android.webkit.WebChromeClient;
+import android.webkit.WebView;
 
 public interface BrowserController {
     void updateAutoComplete();
@@ -14,5 +16,6 @@ public interface BrowserController {
     void onShowCustomView(View view, WebChromeClient.CustomViewCallback callback);
     void onLongPress(String url);
     void hideOverview ();
+    void onCreateView(WebView view, Message resultMsg);
     boolean onHideCustomView();
 }