+++ app/src/main/java/de/baumann/browser/browser/NinjaWebChromeClient.java
@@ -4,6 +4,7 @@ import android.app.Activity;
 import android.content.Context;
 import android.content.Intent;
 import android.net.Uri;
+import android.os.Message;
 import android.view.View;
 import android.webkit.*;
 
@@ -20,13 +21,9 @@ public class NinjaWebChromeClient extends WebChromeClient {
     }
 
     @Override
-    public boolean onCreateWindow(WebView view, boolean dialog, boolean userGesture, android.os.Message resultMsg) {
-        WebView.HitTestResult result = view.getHitTestResult();
-        String data = result.getExtra();
-        Context context = view.getContext();
-        Intent browserIntent = new Intent(Intent.ACTION_VIEW, Uri.parse(data));
-        context.startActivity(browserIntent);
-        return false;
+    public boolean onCreateWindow(WebView view, boolean isDialog, boolean isUserGesture, Message resultMsg) {
+        ninjaWebView.getBrowserController().onCreateView(view, resultMsg);
+        return isUserGesture;
     }
 
     @Override